+++ include/miniRT_parsing.h
@@ -26,27 +26,32 @@ typedef struct s_object_amount
 
 }	t_obj_n;
 
+//file reading
 char	**read_file(char *filename);
 int		valid_category(char **content, int objs_num[6]);
+
+//general parsers
 int		parse_ambient(t_scene *scene, char **content);
 int		parse_camera(t_scene *scene, char **content);
 int		parse_light(t_scene *scene, char **content);
-char	*skip_white(char *str);
-char	*find_content_str(char **content, char *item);
-double	get_numbers(char *str, int type);
-int		save_vector(t_scene *scene, char *content_str, int obj_type, int i);
-int		save_cords(t_scene *scene, char *content_str, int obj_type, int i);
-int		extract_rgb(t_rgb *colors, char *content);
-void	extract_error(int object);
 int		parse_sphere(t_scene *scene, char **content, int n_objs);
 int		parse_plane(t_scene *scene, char **content, int n_objs);
 int		parse_cylinder(t_scene *scene, char **content, int n_objs);
+
+//extracting values or elements
+double	get_numbers(char *str, int type);
+int		save_cords(t_scene *scene, char *content_str, int obj_type, int i);
 int		extract_sphere(t_scene *scene, char **content, int i);
 int		extract_plane(t_scene *scene, char **content, int i);
 int		extract_cylinder(t_scene *scene, char **content, int i);
-void	convert_tabs(char *str);
-int		erreur_dictateur(char **array, int obj_type);
 int		extract_vector(t_vector *vector, char *content);
 int		extract_cords(t_coordinates *cords, char *content);
+int		extract_rgb(t_rgb *colors, char *content);
+
+//utils
+void	convert_tabs(char *str);
+int		erreur_dictateur(char **array, int obj_type);
+char	*find_content_str(char **content, char *item);
+void	extract_error(int object);
 
 #endif //MINIRT_PARSING_H
